v0.7.5: removed unused imports, semicolons, and some unnecessary casts

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/windows/WndTabbed.java
@@ -85,7 +85,7 @@ public class WndTabbed extends Window {
 			width + chrome.marginHor(),
 			height + chrome.marginVer() );
 		
-		camera.resize( (int)chrome.width, (int)(chrome.marginTop() + height + tabHeight()) );
+		camera.resize( (int)chrome.width, chrome.marginTop() + height + tabHeight() );
 		camera.x = (int)(Game.width - camera.screenWidth()) / 2;
 		camera.y = (int)(Game.height - camera.screenHeight()) / 2;
 		camera.y += yOffset * camera.zoom;
